+++ lib/src/enums/_index.dart
@@ -1,3 +1,4 @@
 export 'prf_error_severity.dart';
 export 'prf_error_type.dart';
 export 'prf_failure.dart';
+export 'prf_snackbar_type.dart';